Tests refactored.
